pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["Tatiana Chepurnaya <tanyusha_tokor@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.1"
drf-yasg = "^1.21.7"
setuptools = "^74.1.2"
django-cors-headers = "^4.4.0"
requests = "^2.32.3"
celery = "^5.4.0"
redis = "^5.0.8"
django-celery-beat = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
